Added debug plot to understand sampler behavior

+++ generate.py
@@ -434,8 +434,8 @@ def main(network_pkl, outdir, subdirs, seeds, class_idx, max_batch_size, device=
                 else:
                     PIL.Image.fromarray(image_np, 'RGB').save(image_path)
 
-            lpips_vals.append(loss_fn_vgg(images, x).mean().detach().cpu().numpy())
-            psnr_vals.append(peak_signal_noise_ratio(images, x).mean().detach().cpu().numpy())
+            lpips_vals.append(loss_fn_vgg(images.float(), x.float()).mean().detach().cpu().numpy())
+            psnr_vals.append(peak_signal_noise_ratio(images.float(), x.float()).mean().detach().cpu().numpy())
 
         print(f'Avg. LPIPS: {np.mean(lpips_vals)} +/- {np.std(lpips_vals) / len(lpips_vals)}')
         print(f'Avg. PSNR: {np.mean(psnr_vals)} +/- {np.std(psnr_vals) / len(psnr_vals)}')
